added mini result chart module. #495
